// plugins/release-manager-as-a-service/src/types/types.ts
interface ComponentConfig<Args = void> {
  successCb?: (args: Args) => Promise<void> | void;
  omit?: boolean;
}

export interface ComponentConfigCreateRcSuccessCbArgs {
  gitHubReleaseUrl: string;
  gitHubReleaseName: string;
  comparisonUrl: string;
  previousTag?: string;
  createdTag: string;
}
export type ComponentConfigCreateRc = ComponentConfig<ComponentConfigCreateRcSuccessCbArgs>;

export interface ComponentConfigPromoteRcSuccessCbArgs {
  gitHubReleaseUrl: string;
  gitHubReleaseName: string;
  previousTagUrl: string;
  previousTag: string;
  updatedTagUrl: string;
  updatedTag: string;
}
export type ComponentConfigPromoteRc = ComponentConfig<ComponentConfigPromoteRcSuccessCbArgs>;

export interface ComponentConfigPatchSuccessCbArgs {
  updatedReleaseUrl: string;
  updatedReleaseName: string;
  previousTag: string;
  patchedTag: string;
  patchCommitUrl: string;
  patchCommitMessage: string;
}
export type ComponentConfigPatch = ComponentConfig<ComponentConfigPatchSuccessCbArgs>;

export type SetRefetch = React.Dispatch<React.SetStateAction<number>>;

export interface ResponseStep {
  message: string | React.ReactNode;
  secondaryMessage?: string | React.ReactNode;
  link?: string;
  icon?: 'success' | 'failure';
}
// ...
export interface GhGetReleaseResponse {
  /** @example 'https://api.github.com/repos/octocat/Hello-World/releases/1'; */
  url: string;
  /** @example 'https://github.com/octocat/Hello-World/releases/v1.0.0'; */
  html_url: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/releases/1/assets'; */
  assets_url: string;
  /** @example 'https://uploads.github.com/repos/octocat/Hello-World/releases/1/assets{?name,label}'; */
  upload_url: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/tarball/v1.0.0'; */
  tarball_url: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/zipball/v1.0.0'; */
  zipball_url: string;
  id: number;
  /** @example 'MDc6UmVsZWFzZTE='; */
  node_id: string;
  /** @example 'v1.0.0'; */
  tag_name: string;
  /** @example 'master'; */
  target_commitish: string;
  /** @example 'v1.0.0'; */
  name: string;
  /** @example 'Description of the release'; */
  body: string;
  draft: boolean;
  prerelease: boolean;
  /** @example '2013-02-27T19:35:32Z'; */
  created_at: string;
  /** @example '2013-02-27T19:35:32Z'; */
  published_at: string;
  author: {
    /** @example 'octocat'; */
    login: string;
    id: number;
    /** @example 'MDQ6VXNlcjE='; */
    node_id: string;
    /** @example 'https://github.com/images/error/octocat_happy.gif'; */
    avatar_url: string;
    /** @example ''; */
    gravatar_id: string;
    /** @example 'https://api.github.com/users/octocat'; */
    url: string;
    /** @example 'https://github.com/octocat'; */
    html_url: string;
    /** @example 'https://api.github.com/users/octocat/followers'; */
    followers_url: string;
    /** @example 'https://api.github.com/users/octocat/following{/other_user}'; */
    following_url: string;
    /** @example 'https://api.github.com/users/octocat/gists{/gist_id}'; */
    gists_url: string;
    /** @example 'https://api.github.com/users/octocat/starred{/owner}{/repo}'; */
    starred_url: string;
    /** @example 'https://api.github.com/users/octocat/subscriptions'; */
    subscriptions_url: string;
    /** @example 'https://api.github.com/users/octocat/orgs'; */
    organizations_url: string;
    /** @example 'https://api.github.com/users/octocat/repos'; */
    repos_url: string;
    /** @example 'https://api.github.com/users/octocat/events{/privacy}'; */
    events_url: string;
    /** @example 'https://api.github.com/users/octocat/received_events'; */
    received_events_url: string;
    /** @example 'User'; */
    type: string;
    site_admin: boolean;
  };
  assets: [
    {
      /** @example 'https://api.github.com/repos/octocat/Hello-World/releases/assets/1'; */
      url: string;
      /** @example 'https://github.com/octocat/Hello-World/releases/download/v1.0.0/example.zip'; */
      browser_download_url: string;
      id: number;
      /** @example 'MDEyOlJlbGVhc2VBc3NldDE='; */
      node_id: string;
      /** @example 'example.zip'; */
      name: string;
      /** @example 'short description'; */
      label: string;
      /** @example 'uploaded'; */
      state: string;
      /** @example 'application/zip'; */
      content_type: string;
      size: number;
      download_count: number;
      /** @example '2013-02-27T19:35:32Z'; */
      created_at: string;
      /** @example '2013-02-27T19:35:32Z'; */
      updated_at: string;
      uploader: {
        /** @example 'octocat'; */
        login: string;
        id: number;
        /** @example 'MDQ6VXNlcjE='; */
        node_id: string;
        /** @example 'https://github.com/images/error/octocat_happy.gif'; */
        avatar_url: string;
        /** @example ''; */
        gravatar_id: string;
        /** @example 'https://api.github.com/users/octocat'; */
        url: string;
        /** @example 'https://github.com/octocat'; */
        html_url: string;
        /** @example 'https://api.github.com/users/octocat/followers'; */
        followers_url: string;
        /** @example 'https://api.github.com/users/octocat/following{/other_user}'; */
        following_url: string;
        /** @example 'https://api.github.com/users/octocat/gists{/gist_id}'; */
        gists_url: string;
        /** @example 'https://api.github.com/users/octocat/starred{/owner}{/repo}'; */
        starred_url: string;
        /** @example 'https://api.github.com/users/octocat/subscriptions'; */
        subscriptions_url: string;
        /** @example 'https://api.github.com/users/octocat/orgs'; */
        organizations_url: string;
        /** @example 'https://api.github.com/users/octocat/repos'; */
        repos_url: string;
        /** @example 'https://api.github.com/users/octocat/events{/privacy}'; */
        events_url: string;
        /** @example 'https://api.github.com/users/octocat/received_events'; */
        received_events_url: string;
        /** @example 'User'; */
        type: string;
        site_admin: boolean;
      };
    },
  ];
}